function combining gene lists, alternative for Venn diagram

This function combines multiple gene lists together and summarizes occurrence of the string within a list. Gene lists which will be combined should be a data frame (inputDF) of multiple gene/miRNAslists/or other stings. Gene list in inputDF don't need to have equal length. Output provides logical information if gene (or other string) was present on given list. After using this function you can try to use clusterizer_OneR() on first and last column.

lists.combiner(inputDF)

Arguments

inputDF
inputDF is a data frame in which each column is a different gene list. Gene lists in inputDF don't need to have equal length or be deduplicated. inputDF need to have headings. It can be imported from txt file

Value

Output provides logical information if gene (or other string) was present on given list, and count of occurrences for each gene.

Examples

#example dataframe with gene lists
options(stringsAsFactors = FALSE)
GeneList1=c('PTGER3','FGFR1','ANO5','SGCD','CCKBR',
'PTGFR','NELL1','TBC1D8B','TGFB3')
GeneList2=c('CDH24','PAX1','PTGER3','TBC1D8B','TGFB3',
'PTGER3','NR2C2','NR2C2')
GeneList3=c('PTGER3','FGFR1','ONECUT2','NEBL','SNTB2',
'USP9Y','KAT6A','CRIM1',
'IGSF10','PRKDC','RAD23B','BRD1','PTPRM','PTGIS','RGS5',
'XIRP1','SPSB4')
GeneList4=c('PTGER3','FGFR1','GLIPR1','BCR','SMARCA2','MSMO1','FGF13','KALRN')
x<- list(GeneList1,     GeneList2, GeneList3,   GeneList4)
inputDF<- cbind_filler  (x)

#> [1] "data  1"
#> [1] "data  2"
#> [1] "data  3"
#> [1] "data  4"

names(inputDF)<-c('GeneList1',  'GeneList2', 'GeneList3',       'GeneList4' )
inputDF

#>    GeneList1 GeneList2 GeneList3 GeneList4
#> 1     PTGER3     CDH24    PTGER3    PTGER3
#> 2      FGFR1      PAX1     FGFR1     FGFR1
#> 3       ANO5    PTGER3   ONECUT2    GLIPR1
#> 4       SGCD   TBC1D8B      NEBL       BCR
#> 5      CCKBR     TGFB3     SNTB2   SMARCA2
#> 6      PTGFR    PTGER3     USP9Y     MSMO1
#> 7      NELL1     NR2C2     KAT6A     FGF13
#> 8    TBC1D8B     NR2C2     CRIM1     KALRN
#> 9      TGFB3      <NA>    IGSF10      <NA>
#> 10      <NA>      <NA>     PRKDC      <NA>
#> 11      <NA>      <NA>    RAD23B      <NA>
#> 12      <NA>      <NA>      BRD1      <NA>
#> 13      <NA>      <NA>     PTPRM      <NA>
#> 14      <NA>      <NA>     PTGIS      <NA>
#> 15      <NA>      <NA>      RGS5      <NA>
#> 16      <NA>      <NA>     XIRP1      <NA>
#> 17      <NA>      <NA>     SPSB4      <NA>

#run function
output<- wizbionet::lists.combiner(inputDF)

#> [1] "gene lists in inputDF don't need to have equal lenght or be deduplicated"
#> [1] "done :)"

head(output, n=3)

#>     input GeneList1 GeneList2 GeneList3 GeneList4 count
#> 1  PTGER3      TRUE      TRUE      TRUE      TRUE     4
#> 2   FGFR1      TRUE     FALSE      TRUE      TRUE     3
#> 8 TBC1D8B      TRUE      TRUE     FALSE     FALSE     2