Prioritizes microRNA-target interactions from the the multiMiR Package

This function retrieves miRNA-target interactions and and identify miRNAs and genes with highest number of analyzed interactors.

topmiRNA_toptarget(DEmir_up,DEgenes_down,DEmir_down,
DEgenes_up, multimir_args, mirna_type)

Arguments

DEmir_up
vector with up-regulated miRNAs. To see ID examples see help(get_multimir) and field mirna= . If you don't want to define miRNA put NULL than multimir will analyze all possible pairs

DEgenes_down
vector with down-regulated genes. To see ID examples examples see help(get_multimir) and field target= . If you don't want to define targets put NULL than multimir will analyze all possible pairs

DEmir_down
vector with down-regulated miRNAs

DEgenes_up
vector with up-regulated genes

mirna_type
"mature_mir" or "pre_mir" mirna_type will be used for deduplication and data aggregation

multimir_args
#This parameter is a component of the multiMiR::get_multimir function . You can see description using command: help(get_multimir) #You can modify all components using multimir_args<- as.list(args(multiMiR::get_multimir))

#Important: Don't add mirna= and target= fields they are already included as DEmir_up,DEgenes_down,DEmir_down, DEgenes_up!

Value

This function generates a list with three data frames 1) multimir_output - is a data frame with results from get_multimir function. It provides information about up and down-regulation of the pairs miRNA-targets. It can by used for constructing interaction network in the cytoscape

2) top_miR - is a data frame with aggregated and prioritized results from get_multimir function showing number of genes associated with pre miRNAs It has columns with name clus_... providing logical information if gene was in top 2 clusters (cl1 and cl2) ~top 20 percents and column clusNR_... providing information in which cluster the gene was present (cl1,cl2,cl3,cl4).

2) top_gene - is a data frame with aggregated and prioritized results from get_multimir function showing number of genes associated with analyzed targets. It also has columns with information if gene if in top cluster

References

Ru Y, Kechris KJ, Tabakoff B, Hoffman P, Radcliffe RA, Bowler R, Mahaffey S, Rossi S, Calin GA, Bemis L, Theodorescu D (2014). “The multiMiR R package and database: integration of microRNA–target interactions along with their disease and drug associations.” Nucleic Acids Research, 42(17), e133. doi: 10.1093/nar/gku631, http://dx.doi.org/10.1093/nar/gku631.

Ru Y, Mulvahill M, Mahaffey S, Kechris K. multiMiR: Integration of multiple microRNA-target databases with their disease and drug associations. https://github.com/KechrisLab/multiMiR.

Examples

#Example###

#set parameters
DEmir_up<-c('hsa-miR-150-5p','hsa-miR-448-5p','hsa-miR-448-3p',
'hsa-miR-493-5p','hsa-miR-493-3p') # example DE miRNAs
DEgenes_down<-c('5797','8826','7994','2775','7182','79647','5733',
'158158','9480','8626','50636') # example DE genes
DEmir_down<-c('hsa-miR-4731-5p','hsa-miR-541-3p','hsa-miR-449b-5p','hsa-miR-541-5p')
DEgenes_up<-c('203859','4745','4916','126298','2258','8464','55917','23450','29767')
mirna_type<-"pre_mir" # "mature_mir"
multimir_args= list(url = NULL,
                    org = "hsa",
                    table = "all",
                    predicted.cutoff = 10,
                    predicted.cutoff.type = "p",
                    predicted.site = "conserved"
)

#execute function
output<- wizbionet::topmiRNA_toptarget(DEmir_up,DEgenes_down,
DEmir_down, DEgenes_up, multimir_args,mirna_type)

#> Loading required package: multiMiR

#> Welcome to multiMiR.
#> 
#> multiMiR database URL has been set to the 
#> default value: http://multimir.org/
#> 
#> Database Version: 2.3.0  Updated: 2020-04-15

#> [1] "mirna_type can be \"mature_mir\" or \"premir\", mirna_type will be used for data aggregation "
#> [1] "see all multimir arguments using: as.list(args(multiMiR::get.multimir))"
#> [1] "don't define parameters mirna= and target=, you already defined it in DEmir_up=,DEgenes_down=,DEmir_down=, DEgenes_up= "
#> [1] "multimiR updated: 2020-04-15"

#> Now using database version: 2.3.0

#> [1] "screening DEmir_up and DEgenes_down"
#> Searching mirecords ...
#> Searching mirtarbase ...
#> Searching tarbase ...
#> Searching diana_microt ...
#> Searching elmmo ...
#> Searching microcosm ...
#> Searching miranda ...
#> Searching mirdb ...
#> Searching pictar ...
#> Searching pita ...
#> Searching targetscan ...
#> Searching mir2disease ...
#> Searching pharmaco_mir ...
#> Searching phenomir ...

#> Joining, by = c("database", "mature_mirna_acc", "mature_mirna_id", "target_symbol", "target_entrez", "target_ensembl", "type")

#> Joining, by = c("database", "mature_mirna_acc", "mature_mirna_id", "target_symbol", "target_entrez", "target_ensembl", "type")

#> [1] "screening DEmir_down and DEgenes_up"
#> Searching mirecords ...
#> Searching mirtarbase ...
#> Searching tarbase ...
#> Searching diana_microt ...
#> Searching elmmo ...
#> Searching microcosm ...
#> Searching miranda ...
#> Searching mirdb ...
#> Searching pictar ...
#> Searching pita ...
#> Searching targetscan ...
#> Searching mir2disease ...
#> Searching pharmaco_mir ...
#> Searching phenomir ...

#> Joining, by = c("database", "mature_mirna_acc", "mature_mirna_id", "target_symbol", "target_entrez", "target_ensembl", "type")

#> Joining, by = c("database", "mature_mirna_acc", "mature_mirna_id", "target_symbol", "target_entrez", "target_ensembl", "type")

#> [1] "aggregating data to find top miRNA"
#> [1] "inputDF- dataframe for aggegation and counting (A&C)"
#> [1] "col_names- column names which will be A&C"
#> [1] "col_collapse- landmark column for A&C columns"
#> [1] "rows_collapse- content of the column  landmark column all_of(col_collapse) as vector "
#> [1] "control_col- additional landmark column"
#> [1] "hsa-miR-150 count: 1 mature_mirna_acc"
#> [1] "hsa-miR-150 count: 1 mature_mirna_id"
#> [1] "hsa-miR-150 count: 5 target_symbol"
#> [1] "hsa-miR-150 count: 5 target_entrez"
#> [1] "hsa-miR-150 count: 5 target_ensembl"
#> [1] "hsa-miR-150 count: 2 type"
#> [1] "hsa-miR-150 count: 1 multimir"
#> [1] "hsa-miR-493 count: 2 mature_mirna_acc"
#> [1] "hsa-miR-493 count: 2 mature_mirna_id"
#> [1] "hsa-miR-493 count: 7 target_symbol"
#> [1] "hsa-miR-493 count: 7 target_entrez"
#> [1] "hsa-miR-493 count: 7 target_ensembl"
#> [1] "hsa-miR-493 count: 2 type"
#> [1] "hsa-miR-493 count: 1 multimir"
#> [1] "hsa-miR-541 count: 2 mature_mirna_acc"
#> [1] "hsa-miR-541 count: 2 mature_mirna_id"
#> [1] "hsa-miR-541 count: 3 target_symbol"
#> [1] "hsa-miR-541 count: 3 target_entrez"
#> [1] "hsa-miR-541 count: 3 target_ensembl"
#> [1] "hsa-miR-541 count: 2 type"
#> [1] "hsa-miR-541 count: 1 multimir"
#> [1] "hsa-miR-449b count: 1 mature_mirna_acc"
#> [1] "hsa-miR-449b count: 1 mature_mirna_id"
#> [1] "hsa-miR-449b count: 5 target_symbol"
#> [1] "hsa-miR-449b count: 5 target_entrez"
#> [1] "hsa-miR-449b count: 5 target_ensembl"
#> [1] "hsa-miR-449b count: 2 type"
#> [1] "hsa-miR-449b count: 1 multimir"
#> [1] "hsa-miR-4731 count: 1 mature_mirna_acc"
#> [1] "hsa-miR-4731 count: 1 mature_mirna_id"
#> [1] "hsa-miR-4731 count: 2 target_symbol"
#> [1] "hsa-miR-4731 count: 2 target_entrez"
#> [1] "hsa-miR-4731 count: 2 target_ensembl"
#> [1] "hsa-miR-4731 count: 2 type"
#> [1] "hsa-miR-4731 count: 1 multimir"
#> [1] "OneR clustering data to find top miRNA"
#> [1] "in the column target_symbol_COUNT: cl1=39% 11 rows: cl2=43% 12 rows"
#> [1] "target_symbol_COUNT top25% of values: selected 39% of rows (11 rows) for nr higher than 7"
#> [1] "aggregating data to find top mRNA"
#> [1] "inputDF- dataframe for aggegation and counting (A&C)"
#> [1] "col_names- column names which will be A&C"
#> [1] "col_collapse- landmark column for A&C columns"
#> [1] "rows_collapse- content of the column  landmark column all_of(col_collapse) as vector "
#> [1] "control_col- additional landmark column"
#> [1] "ANO7 count: 1 target_entrez"
#> [1] "ANO7 count: 1 target_ensembl"
#> [1] "ANO7 count: 2 mature_mirna_acc"
#> [1] "ANO7 count: 2 mature_mirna_id"
#> [1] "ANO7 count: 2 premir"
#> [1] "ANO7 count: 2 type"
#> [1] "ANO7 count: 1 multimir"
#> [1] "AKIRIN1 count: 1 target_entrez"
#> [1] "AKIRIN1 count: 1 target_ensembl"
#> [1] "AKIRIN1 count: 1 mature_mirna_acc"
#> [1] "AKIRIN1 count: 1 mature_mirna_id"
#> [1] "AKIRIN1 count: 1 premir"
#> [1] "AKIRIN1 count: 1 type"
#> [1] "AKIRIN1 count: 1 multimir"
#> [1] "IQGAP1 count: 1 target_entrez"
#> [1] "IQGAP1 count: 1 target_ensembl"
#> [1] "IQGAP1 count: 2 mature_mirna_acc"
#> [1] "IQGAP1 count: 2 mature_mirna_id"
#> [1] "IQGAP1 count: 2 premir"
#> [1] "IQGAP1 count: 1 type"
#> [1] "IQGAP1 count: 1 multimir"
#> [1] "KAT6A count: 1 target_entrez"
#> [1] "KAT6A count: 1 target_ensembl"
#> [1] "KAT6A count: 2 mature_mirna_acc"
#> [1] "KAT6A count: 2 mature_mirna_id"
#> [1] "KAT6A count: 2 premir"
#> [1] "KAT6A count: 1 type"
#> [1] "KAT6A count: 1 multimir"
#> [1] "ONECUT2 count: 1 target_entrez"
#> [1] "ONECUT2 count: 1 target_ensembl"
#> [1] "ONECUT2 count: 2 mature_mirna_acc"
#> [1] "ONECUT2 count: 2 mature_mirna_id"
#> [1] "ONECUT2 count: 2 premir"
#> [1] "ONECUT2 count: 1 type"
#> [1] "ONECUT2 count: 1 multimir"
#> [1] "TP63 count: 1 target_entrez"
#> [1] "TP63 count: 1 target_ensembl"
#> [1] "TP63 count: 1 mature_mirna_acc"
#> [1] "TP63 count: 1 mature_mirna_id"
#> [1] "TP63 count: 1 premir"
#> [1] "TP63 count: 1 type"
#> [1] "TP63 count: 1 multimir"
#> [1] "NR2C2 count: 1 target_entrez"
#> [1] "NR2C2 count: 1 target_ensembl"
#> [1] "NR2C2 count: 1 mature_mirna_acc"
#> [1] "NR2C2 count: 1 mature_mirna_id"
#> [1] "NR2C2 count: 1 premir"
#> [1] "NR2C2 count: 1 type"
#> [1] "NR2C2 count: 1 multimir"
#> [1] "GNAO1 count: 1 target_entrez"
#> [1] "GNAO1 count: 1 target_ensembl"
#> [1] "GNAO1 count: 1 mature_mirna_acc"
#> [1] "GNAO1 count: 1 mature_mirna_id"
#> [1] "GNAO1 count: 1 premir"
#> [1] "GNAO1 count: 1 type"
#> [1] "GNAO1 count: 1 multimir"
#> [1] "SF3B3 count: 1 target_entrez"
#> [1] "SF3B3 count: 1 target_ensembl"
#> [1] "SF3B3 count: 3 mature_mirna_acc"
#> [1] "SF3B3 count: 3 mature_mirna_id"
#> [1] "SF3B3 count: 3 premir"
#> [1] "SF3B3 count: 1 type"
#> [1] "SF3B3 count: 1 multimir"
#> [1] "IRGQ count: 1 target_entrez"
#> [1] "IRGQ count: 1 target_ensembl"
#> [1] "IRGQ count: 2 mature_mirna_acc"
#> [1] "IRGQ count: 2 mature_mirna_id"
#> [1] "IRGQ count: 2 premir"
#> [1] "IRGQ count: 1 type"
#> [1] "IRGQ count: 1 multimir"
#> [1] "TMOD2 count: 1 target_entrez"
#> [1] "TMOD2 count: 1 target_ensembl"
#> [1] "TMOD2 count: 1 mature_mirna_acc"
#> [1] "TMOD2 count: 1 mature_mirna_id"
#> [1] "TMOD2 count: 1 premir"
#> [1] "TMOD2 count: 1 type"
#> [1] "TMOD2 count: 1 multimir"
#> [1] "NTRK3 count: 1 target_entrez"
#> [1] "NTRK3 count: 1 target_ensembl"
#> [1] "NTRK3 count: 2 mature_mirna_acc"
#> [1] "NTRK3 count: 2 mature_mirna_id"
#> [1] "NTRK3 count: 2 premir"
#> [1] "NTRK3 count: 1 type"
#> [1] "NTRK3 count: 1 multimir"
#> [1] "SUPT3H count: 1 target_entrez"
#> [1] "SUPT3H count: 1 target_ensembl"
#> [1] "SUPT3H count: 1 mature_mirna_acc"
#> [1] "SUPT3H count: 1 mature_mirna_id"
#> [1] "SUPT3H count: 1 premir"
#> [1] "SUPT3H count: 1 type"
#> [1] "SUPT3H count: 1 multimir"
#> [1] "CTTNBP2NL count: 1 target_entrez"
#> [1] "CTTNBP2NL count: 1 target_ensembl"
#> [1] "CTTNBP2NL count: 1 mature_mirna_acc"
#> [1] "CTTNBP2NL count: 1 mature_mirna_id"
#> [1] "CTTNBP2NL count: 1 premir"
#> [1] "CTTNBP2NL count: 1 type"
#> [1] "CTTNBP2NL count: 1 multimir"
#> [1] "OneR clustering data to find top mRNA"
#> [1] "in the column mature_mirna_id_COUNT: cl1=11% 3 rows: cl2=54% 15 rows"
#> [1] "in the column premir_COUNT: cl1=11% 3 rows: cl2=54% 15 rows"
#> [1] "mature_mirna_id_COUNT top25% of values: selected 64% of rows (18 rows) for nr higher than 2"
#> [1] "premir_COUNT top25% of values: selected 64% of rows (18 rows) for nr higher than 2"
#> [1] "                   Please cite creators of multiMiR doi: 10.1093/nar/gku631"
#> [1] "                   Please cite creator of this package: Wicik Z."